+++ Encryption/EncryptedDataBlock.cs
@@ -1,0 +1,10 @@
+namespace N17Solutions.Semaphore.Encryption
+{
+    public class EncryptedDataBlock
+    {
+        public string EncryptedData { get; set; }
+        public string DigitalSignature { get; set; }
+        public string InitialisationVector { get; set; }
+        public string AesKey { get; set; }
+    }
+}